requirements.txt
Django==1.9.4
django-rest-auth==0.7.0
django-simple-history==1.8.1
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.7
ipaddress==1.0.16
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
wsgiref==0.1.2
